[show] Add subcommand to display the status of auto-restart feature for each container (#798)
Since we introduced the auto-restart feature for each container, we need add a show subcommand to display the current status of auto-restart feature for all containers or a specific container.

- How I did it
We define a function named autorestart to show the status of this features. This function will accept an option parameter which is the container name. If this parameter is not specified, this function will by default show the status of all containers. Otherwise it will show the status of specific container.

Signed-off-by: Yong Zhao <yozhao@microsoft.com>