[config] Remove _get_breakout_cfg_file_name helper function (#1069)
This is part of a larger clean up of functions which return the port_config_file which reside outside of
sonic_py_common packages. This is motivated due to redundancy of code as well as making it easier to change
the default behaviour of helper utility function port_config_file unified and in one place.