Remove unnecessary conversions to list() and calls to dict.keys() (#1243)
After conversion to Python 3 using the 2to3 tool, all calls to `dict.keys()` and `dict.values()` were converted to lists. In the instances where the returned view objects from these calls are iterated over, there is no need to convert to a list, so I removed those.

Also, when iterating over the keys of a dictionary, one need not call `dict.keys()`, one can simply iterate over the dictionary as `for key in dict:`, so I cleaned those up as well.