[config vlan]: Remove `-t` flag from docker exec command (#1317)
Remove -t flag from the docker exec command used to restart SWSS after updating the proxy ARP config. Currently with the flag, this command cannot be executed as part of a script (e.g. during a unit test in sonic-mgmt) since Docker is expecting the input device to be a TTY.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>