Fix issues in clear_qos (#2122)
- What I did
Fix issues in clear_qos

1. Remove tables BUFFER_PORT_INGRESS_PROFILE_LIST and BUFFER_PORT_EGRESS_PROFILE_LIST
When clear_qos is executed, all QoS- and buffer-related tables should be removed. Currently, both tables are missed. This is to fix it.
2. Adjust the order in which the buffer and QoS tables are removed
Remove the tables whose entries depend on other tables first and then the tables whose entries are referenced by other tables.
Any object can be removed from SAI only if there is no object referencing it. There are retrying and dependency-tracking mechanisms to guarantee it in orchagent.
If CLI operates the table in an order where the referencing tables removed first before referenced tables, it is possible to reduce the probability to retry theoretically.
The order does not guarantee it but adjusting the order is almost zero-cost so we would like to do it.

- How I did it

- How to verify it

Signed-off-by: Stephen Sun stephens@nvidia.com