[generate dump]Added error message when saisdkdump fails (#2356)
- What I did
Added error message when saisdkdump fails and cont to gather the rest. This is done to provide more readable information to the user when it cannot be avaialble (syncd for example is not running, during restart, etc).

- How I did it
Checked error code and print log

- How to verify it
Simulate saisdkdump error and verify the console output.