[subinterface]Added additional checks in portchannel and subinterface commands (#2345) (#2371)
Backport of #2345

- What I did
Added additional checks in subinterface and portchannel commands so they don't conflict. Without the checks, a subinterface could be created on a portchannel member and vice versa which will lead to SAI failure followed by orchagent crash.

- How I did it
Added checks in relevant config functions.

- How to verify it
Added UT to verify it.