Do not configure physical attributes on port channels in portconfig (#2456)
- What I did
portconfig is a utility designed to configure various attributes on a physical port or a port channel.
It should reject the operation that a user wants to configure an attribute that will be inserted into CONFIG_DB.PORT table on a port channel. Otherwise, it will generate an entry for the port channel, which causes orchagent to be aborted.

Eg. the following commands will cause an error

admin@sonic:~$ sudo config portchannel add PortChannel0001
admin@sonic:~$ sudo config interface speed PortChannel0001 100000

After the fix, configuring physical attributes on a port channel will be rejected

- How I did it

- How to verify it
Unit test

Signed-off-by: Stephen Sun <stephens@nvidia.com>