Support to display the SONiC OS Version in the command show version (#2787)
What I did
Support to display the SONiC OS Version in the command show version.
It will be used to display the version info in the SONiC command "show version". The version is used to do the FIPS certification. We do not do the FIPS certification on a specific release, but on the SONiC OS Version.

SONiC Software Version: SONiC.master-13812.218661-7d94c0c28
SONiC OS Version: 11
Distribution: Debian 11.6
Kernel: 5.10.0-18-2-amd64
How I did it
The device info is in sonic-net/sonic-buildimage, see PR: sonic-net/sonic-buildimage#14601
The submodule change can be merged to sonic-buildimage, after the PR 14601 merged.

How to verify it