[sonic-package-manager] add support for multiple CLI plugin files (#2753)
#### What I did

Use case for some application extensions we develop is to have multiple features as part of a single container. Therefore we want to have ability to install multiple CLI plugins from a single extension.

#### How I did it

Allowed to specify few CLI plugins in the manifest.

#### How to verify it

Build extension with multiple CLI plugins and install. Verified all plugins are installed.