[GCU]Fix rdma check failure (#2824)
What I did
if substring.isdigit() and re.match(r'^\d{8}$', substring): In this version check, it only applies to this pattern: 20220531.XXX. It will fail the master and other build check. For example, master.261851-010dc3957 is a master build_version string. It will fail version validation and return failure.

My thought to remove the check is that GCU is only supported in 202205 and later. Either branch_version >= "20201200" or branch_version >= "20181100" is already satisfied.

How I did it
Remove redundant check in GCU RDMA validator

How to verify it
Unit test