clear: Fix clear queuecounters to also clear VOQ counters (#2879)
The `sonic-clear queuecounters` command does not currently clear VOQ
counters. Fix it and add a unit test to verify that the VOQ counters
were actually cleared.

The unit test changes required some updates to the VOQ counters tests,
because on an actual DUT the VOQ counter uses fully-qualified
interface names (dut|asic|port) whereas the queue counters just use
the port name. The unit tests had kept the names the same but this
meant that the saved counters for VOQ and non-VOQ counters would
overwrite each other when the counters were cleared during the unit
tests. Qualify the interface names used in the VOQ unit tests to avoid
this.